Refactor in prep for gimlet/sidecar/psc variants of mgmt-gateway
No functional changes. Moves usart logic out of main and into
`mgs_handler` (which will become the gimlet-specific MGS handler).
